existing states structuring
