+++ FindCar/FindCar/Bot/States/PassengerStates/FindCarState.cs
@@ -1,14 +1,13 @@
 using System.Threading.Tasks;
 using Telegram.Bot.Types;
 
-namespace FindCar.Bot
+namespace FindCar.Bot.PassengerStates
 {
     public class FindCarState : IBotState
     {
         public async Task OnInit(ChatContext ctx)
         {
-            await ctx.SendText("Створiть заявку за форматом:\n" +
-                               "Звiдки:Куды:");
+            await ctx.SendText("В якому ви місті?");
         }
 
         public async Task<IBotState> HandleMessage(ChatContext ctx, Message message)
