+++ FindCar/FindCar/Bot/States/StartedState.cs
@@ -1,22 +1,14 @@
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
+﻿using System.Threading.Tasks;
+using FindCar.Bot.PassengerStates;
 using Telegram.Bot.Types;
 
 namespace FindCar.Bot
 {
-    public interface IBotState
-    {
-        Task OnInit(ChatContext ctx);
-        Task<IBotState> HandleMessage(ChatContext ctx, Message message);
-    }
-
     public class Started : IBotState
     {
         private const string FindCar = "Я шукаю авто";
         private const string HaveCar = "Я маю авто";
-
-
+        
         public async Task OnInit(ChatContext ctx)
         {
             await ctx.Send("Вкажiть що вы хочете зробити", new []
