+++ FindCar/FindCar/Bot/Store.cs
@@ -1,4 +1,3 @@
-using System;
 using System.Threading.Tasks;
 using MongoDB.Driver;
 
@@ -30,16 +29,21 @@ namespace FindCar.Bot
         
         public Task SaveState<T>(long chatId, T state) where T : IBotState
         {
+            // TODO implement
             //var state = _db.GetCollection<JsonWrap>();
+            return null;
         }
 
         public Task<IBotState> GetState(long chatId)
         {
-            
+            // TODO implement
+            return null;
         }
         
         public Task<Passenger[]> GetPassengers(int seats, string from, string to)
         {
+            // TODO implement
+            return null;
         }
     }
 }
